tests/wpt/tests/html/semantics/links/links-created-by-a-and-area-elements/noopener_case_sensitivity.html

<!DOCTYPE html>
<html>
<head>
  <meta charset=utf-8>
  <meta name="timeout" content="long">
  <meta help="https://html.spec.whatwg.org/multipage/links.html#get-an-element's-noopener">
  <title>Test that "rel" attribute values are treated case-insensitively when getting an elements noopener</title>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
</head>
<body>
  <!-- All of the "rel" keywords contain uppercase letters, which should not cause issues -->
  <a href="support/target_blank_implicit_noopener.html?a1" id="a1" rel="Noopener">Click me</a>
  <a href="support/target_blank_implicit_noopener.html?a6" id="a2" rel="Noreferrer">Click me</a>
  <a href="support/target_blank_implicit_noopener.html?a2" id="a3" rel="Opener" target="_blank">Click me</a>

  <script>

  let tests = [
    { id: "a1", hasOpener: false, name: "Anchor element with rel=Noopener" },
    { id: "a2", hasOpener: false, name: "Anchor element with rel=Noreferrer" },
    { id: "a3", hasOpener: true, name: "Anchor element with base target=_blank with rel=Opener" },
  ];

  tests.forEach(data => {
    async_test(
      test => {
        let bc = new BroadcastChannel(data.id);
        bc.addEventListener("message", test.step_func_done(e => {
          assert_equals(e.data.hasOpener, data.hasOpener);
        }), {once: true});

        document.getElementById(data.id).click();
      }, data.name);
  });
  </script>
</body>
</html>
